dyn major change & dyn update on select & spell prediction #88 initial test & few core functions & bugfix
This commit has major change to dyn

`meta data-dyn-insert-in` and `meta data-dyn-replace-node` has removed.
For now on dny will only perform as `dyn-replace-node` style.
This change has significant simplification on implementation of dyn.
`dyn-insert-in` is similar to having a node with a signle subnode with meta dyn.
So the functionality is not limited.

With this change updating dyn can happen on the fly without the need to restart main process. Which this commit contains impl of update dyn on select of node that has `meta dyn-setdirty-onselect` on the active tree.

The commit contains one or more un-tapped bugs.
1. bugfix for `tree_add_node` with `at` argument
2. `_tree_go_out` unset `spell-branch` when goes out of `spell-branch`