impl of #88 , word prediction for spell-branch & bugfixes
Hello webpack, babel & es6
For adding module "sorted-array-functions" from npm to the app, I did add
webpack and babel to gulpfile. Now `./node_modules/gulp build-dev` is needed for deployment.

Why did i add "sorted-array-functions"?
Well i came to a conclusion that a simple and reliable runtime word prediction
utility is to have an alphabetically sorted array of words with their weights
for guiding the importance of that word.
And this module has one useful optimized function `eq`, For searching in a sort array. With T = O(log(n)) performance

dyn changes
Couple of bugfixes, disable dyn when edit_mode is on.

`meta spell-word="SOMEWORD"` use of this meta in a `spell-branch` will result in replacing the word with incomplete letters selected. This meta is used at dyn `spell-word-prediction`.
`meta spell-finish` can now include `spell-word` or `spell-letter` with it. To include the word or letter and then finish.

`meta data-spell-update-dyn-onchange` This meta tag updates all sub nodes dyn when a change occurs to `spell-branch`. It is needed to be set at `spell-branch` node to have it update prediction dyn nodes.

`meta dyn="spell-word-prediction"` is the word prediction dynamic node.
  It requires `words-file="URL"` which points at words file with json content type. Words within `words.json` should be all in lower case.
  `max-nodes="number"` sets a limit for amount of words to predict
  `spell-finish` can be used to indicate spell process should finish at select

`meta dyn="spell-letter-prediction"` sorts alphabet according to weights of existing words that match or alphabetic order.
  It requires `word-file="URL"`, Which is same as `spell-word-prediction`.
  `alphabet="abcd...."` or ="a,b,c,d..."

minor changes
1. On fatal-error at startup, show menubar + more, So user can have a chance to fix the issue at config if possible
2. bugfix of parse_dom_tree when cnode is heading and meta is inside it