upgrade to 1.0 pasco-data-state & dropbox-sync & recreate cordova project + random fixes
* Upgrade to PascoDataState v1.0

The PascoDataState upgrade is about converting all links in all files
as relative links to the place it has been declared. This upgrade allows
copy of all stored data to other platforms with no need to modify any
content at the time of import or export. Moreover it introduces the
pasco-state.json file which holds the list of all stored files in
pasco with their checksum and file size.

* Dropbox Sync

The sync tab has been added in edit-config page. Everything related
to saving and loading data from dropbox is in there.
It should be easy to use.

* Upgrade to cordova v6

I tried to run cordova on macos Big Sur with no success. So I did end up
re-creating cordova configuration. With that still few files has been
modified in the cordova directory. The same install process can be used
with a new clone of pasco. But I think the cordova-icon generator should
get called before publishing the app.

A custom protocol has been added to cordova ios (pasco://),
It enables the authorization of dropbox.

* Note

Two required environment variables has been introduced in this commit.
They are required when running build scripts.
1. PASCO_BACKEND_URL (default value is 'https://backend.pasco.chat')
2. DROPBOX_APP_KEY (Should be the client_id of the dropbox app)

related to #243